requirements.txt
aniso8601==8.0.0
astroid==2.3.3
bcrypt==3.1.7
blinker==1.4
cffi==1.14.0
click==7.1.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==3.24.1
Flask-Mail==0.9.1
flask-mongoengine==0.9.5
Flask-RESTful==0.3.8
Flask-WTF==0.14.3
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mongoengine==0.19.1
mypy==0.770
mypy-extensions==0.4.3
pycparser==2.20
PyJWT==1.7.1
pylint==2.4.4
pymongo==3.10.1
python-dotenv==0.13.0
pytz==2019.3
six==1.14.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
Werkzeug==1.0.1
wrapt==1.11.2
WTForms==2.2.1
